Clean up unnecessary code for the calendar
Had some things from the calendar that were not necessary.
