+++ Assets/Scenes/SampleScene.unity
@@ -2293,8 +2293,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 1173785612}
-        m_TargetAssemblyTypeName: YearMonthUpDown, Assembly-CSharp
+      - m_Target: {fileID: 1173785611}
+        m_TargetAssemblyTypeName: Calendar, Assembly-CSharp
         m_MethodName: YearUp
         m_Mode: 1
         m_Arguments:
@@ -4753,7 +4753,7 @@ MonoBehaviour:
   m_TargetGraphic: {fileID: 799638377}
   m_HandleRect: {fileID: 799638376}
   m_Direction: 2
-  m_Value: 0.99999017
+  m_Value: 0.99999803
   m_Size: 1
   m_NumberOfSteps: 0
   m_OnValueChanged:
@@ -8767,8 +8767,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 1173785612}
-        m_TargetAssemblyTypeName: YearMonthUpDown, Assembly-CSharp
+      - m_Target: {fileID: 1173785611}
+        m_TargetAssemblyTypeName: Calendar, Assembly-CSharp
         m_MethodName: YearDown
         m_Mode: 1
         m_Arguments:
@@ -9058,8 +9058,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 1173785612}
-        m_TargetAssemblyTypeName: YearMonthUpDown, Assembly-CSharp
+      - m_Target: {fileID: 1173785611}
+        m_TargetAssemblyTypeName: Calendar, Assembly-CSharp
         m_MethodName: MonthUp
         m_Mode: 1
         m_Arguments:
@@ -13341,8 +13341,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 1173785612}
-        m_TargetAssemblyTypeName: YearMonthUpDown, Assembly-CSharp
+      - m_Target: {fileID: 1173785611}
+        m_TargetAssemblyTypeName: Calendar, Assembly-CSharp
         m_MethodName: MonthDown
         m_Mode: 1
         m_Arguments:
