--- Assets/Scripts/UI/YearMonthUpDown.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: d21ec44e0b2377c45b2cf4a9751f9858
